requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.0
itsdangerous>=2.1
matplotlib>=3.8
numpy>=1.24
pandas>=2.0
pylatexenc>=2.10
python-dateutil>=2.8
reportlab>=4.0
scikit-learn>=1.5
seaborn>=0.13
Werkzeug>=3.0
Flask-Login>=0.6
Flask-WTF>=1.1
email-validator>=2.0
